backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
web3==6.11.3
eth-account==0.9.0
cryptography==41.0.7
qrcode==7.4.2
geopy==2.4.0
drf-spectacular==0.26.5
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
locust==2.17.0
